Refactor to put different test areas in their own sourcefiles.
